pyproject.toml
[project]
name = "q3logs2kafka"
description = "Send q3 logs as events to kafka"
version = "0.0.1"
authors = [
    { name = "Marco Siebecke", email = "marco@siebecke.se" }
]
dependencies = [
    "click",
    "kafka-python"
]

[project.scripts]
q3logs_reader = "q3logs_reader.main:cli"
event_producer = "event_producer.main:cli"
